move collection into collections folder
